app/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.27.1
numpy==1.24.4
pandas==2.0.3
pytesseract==0.3.13
Pillow==10.1.0
python-multipart==0.0.17
prometheus-client==0.19.0